style(Gausium Connector): Move magic strings to a constants file
