+++ gausium_connector/inorbit_gausium_connector/src/robot/constants.py
@@ -1,0 +1,20 @@
+# SPDX-FileCopyrightText: 2025 InOrbit, Inc.
+#
+# SPDX-License-Identifier: MIT
+
+
+from enum import Enum
+
+
+class WorkType(Enum):
+    """Enum to hold possible robot work types."""
+
+    NAVIGATING = "NAVIGATING"
+    EXECUTE_TASK = "EXECUTE_TASK"
+
+
+class TaskState(Enum):
+    """Enum to hold possible task states."""
+
+    STARTED = "STARTED"
+    PAUSED = "PAUSED"
